requirements.txt
Pillow~=9.0.1
cryptography~=36.0.1
nuitka==1.1.2